requirements.txt
aiosmtplib==5.1.0
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
beautifulsoup4==4.13.5
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.116.1
fastapi-mail==1.6.2
greenlet==3.3.2
h11==0.16.0
httptools==0.7.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
passlib==1.7.4
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.10.1
pydantic_core==2.41.5
PyJWT==2.11.0
PyMySQL==1.1.1
python-dotenv==1.0.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
soupsieve==2.8.3
SQLAlchemy==2.0.43
starlette==0.47.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.35.0
watchfiles==1.1.1
websockets==16.0
